Fixing folder structure
Addressing issues with moving the folders from gitlab to github
